Reorganise test project structure and DI
Reorganises project structure for tests so that they reflect the
directories in the parent. This brings us into line with other Python
projects and reduces confusion.

Changes the DI to move away from patching; this reduces boilerplate and
also simplifies the test injection. We can use a type injection in a
base wrapper to effectively template the injected class. This allows us
to still use the injected class within a context manager whilst getting
rid of patching at the same time.