MAINT: add Results and ResultsQuery classes
Result class is used to store a single query result item. 
It stores the openstack item as well as any forwarded properties.
 
This class is responsible for outputting the value stored as_object or as_props. 

When as_props() is selected, it will concatenate any forwarded properties stored automatically 


ResultContainer class is responsible for storing multiple Result objects and returning the total output to_objects/to_props. 
It is responsible for creating and managing each Result item once a query is run

Currently neither class is being used - Query Executer will make use of  this in future - this PR is being split up so it can be reviewed quicker