MAINT: refactor decom flavor to find servers with missing user info
A few changes to decom flavor email action to fix bugs where if a server with an associated user id was not found via a UserQuery, it would be ignored.

These changes mean that we separate out finding user info from the original query. for each user id we encounter in the original query, we run a separate user query to get the info