Fix matcher
As explained in JIRA ticket https://stfc.atlassian.net/browse/CLDSCRM-2323

The current alertmanager silence scheduler does not silence
all alerts as there are alerts that search for hostname
instead of instance.  It may be worth adding “matchers”
as a parameter to be passed to the schedule_silence method.

Therefore, we replace string "instance" by "hostname" everywhere in the
code.