Add function to reply to JIRA Issues
As it is going to be needed by future automations, we need to enhance the
API to interact with JIRA.
Let's start by adding a function that allows us to reply to an existing
JIRA Issue. The reply can be either an internal comment or a reply to
the customer.