Add minimal code for the Sensor
The Sensor in this automation will be minimal. It will just poll the
JIRA server periodically and check if there are new JIRA Issues to be
processed.

The JIRA Issues to be processed by this automation are easily identified
as they must be in the right state: "Ready For Automation".

Once one or more JIRA Issues have been identified, the Sensor simply
triggers the action, passing as payload data just the unique Issue Key.
The Sensor won't do anything else, the rest of the automation will be
performed by the Action.