Add new email templates
Add 2 new JINJA2 templates (txt and HTML) for emails notifications.
This new templates will be used to send a message to an specific email
address to notify when a CAPI image is about to be decommissioned.
In that case, we will notify the email address registered for the
project(s) currently using that CAPI image.

The YAML file "email_template_schemas" also updated accordingly.