Move function find_servers_with_decom_flavor
Move the function to find servers with a given flavor name (or names) to
the API directory for the query library:

      lib/apis/openstack_query_api

The code to send emails shouldn't need to know about the internal
details of the query library.
Moreover when there is an API directory for that query library.
All code that needs to use it should be implemented as part of the API.

In this commit we implement the following changes:

* move the function to find servers to the query library API

* remove the function to find servers from the workflow
  send_decom_flavor_email.py

* the block of code to raise an Exception if no server is found remains
  in the workflow, that part does not need to be moved to the API

* fix the docstrings after moving the code to the API directory, to
  remove explicit mentions to "decommissioning"

* implement a new unittest file for the new code in the API directory

* fix some incorrect function names in the old unittest code
  (functions had string "user" where it should be "servers")

* fix the unittest code for the workflow