BUG: Revert cpu_cores back to cores from 2bd55786
Quota set value was changed to cpu_cores as part of a descriptive
renaming PR. This key value still needs to be cores as the openstack api
expects cores not cpu_cores. We can keep the action as cpu_cores and
just do this translation later on.

Update the tests as well.